Refactored recommendations.ts to move implementation away from index.tsx page as much as possible
