+++ lib/recommendations.ts
@@ -1,31 +1,50 @@
-import {Customer} from "@commerce/types/customer";
 import {OrderEdge, OrderLineItemEdge} from "@framework/schema";
 import {Product} from "@commerce/types/product";
 import _ from "lodash";
+import {useCustomerOrders} from "@framework/customer";
+import {useSearch} from "../framework/local/product";
+import {useEffect, useState} from "react";
 
-export function collectOrderedBrands(customerOrders: Customer, orderedBrands: Array<string>, setOrderedBrands: (value: (((prevState: Array<string>) => Array<string>) | Array<string>)) => void) {
+export function useRecommendedProduct(locale: string | undefined): Product | undefined {
+  const [orderedBrands, setOrderedBrands] = useState<Array<string>>([])
+  const [recommendedProduct, setRecommendedProduct] = useState<Product>()
+
+  const { data: customerOrders } = useCustomerOrders({ numberOfOrders: 3 }) // Get customer orders
+
+  const { data: recommendedProducts } = useSearch({ // Get products based on random ordered brand
+    brandId: getRandomBrandId(orderedBrands), // If no ordered brands search for placeholder string (No results)
+    locale,
+  })
+
+  useEffect(() => {
   customerOrders?.orders.edges.map((order: OrderEdge) => { // Iterate through orders
     order.node.lineItems.edges.map((item: OrderLineItemEdge) => { // Iterate through ordered items
-      if (item.node.variant?.product.vendor && orderedBrands.indexOf(item.node.variant.product.vendor) === -1) { // If vendor is known and doesn't exist in orderedBrands list yet
-        const vendor = item.node.variant.product.vendor
+      const vendor = item.node.variant?.product.vendor
+      if (vendor && orderedBrands.indexOf(vendor) === -1) { // If vendor is known and doesn't exist in orderedBrands list yet
         setOrderedBrands(previousOrderedBrands => [...previousOrderedBrands, vendor]) // Collect list of ordered brands
       }
     })
   })
-}
+  }, [customerOrders?.orders.edges, orderedBrands])
 
-export function getRecommendedProduct(customerOrders: Customer, recommendedProducts: { products: Product[]; found: boolean; } | undefined, productList: Product[], setProductList: (value: (((prevState: any) => any) | any)) => void) {
-  if (customerOrders?.orders.edges.length) {
-    const recommendedProduct = _.sample(recommendedProducts?.products) // Take random recommended product from brand
-    if (recommendedProduct) {
-      const productListCopy = [...productList]
-      const result = productListCopy.findIndex(product => { return product.id === recommendedProduct?.id }) // Check if recommended product already exists in list of products
-      if (result != -1) { productListCopy.splice(result, 1) } // If it exists remove it
-      recommendedProduct.description += '-recommended-' // Alter description of recommended product to mark it as recommended
-      if (!productListCopy.some(product => product.description?.includes('-recommended-'))) { // Check if no product has been recommended already
-        setProductList([recommendedProduct, ...productListCopy]) // Add recommended product to beginning of list
+  useEffect(() => {
+    if (customerOrders?.orders.edges.length) {
+      const randomProduct = _.sample(recommendedProducts?.products) // Take random recommended product from brand
+      if (randomProduct) {
+        randomProduct.recommended = true // Alter description of recommended product to mark it as recommended
+        setRecommendedProduct(randomProduct)
       }
     }
+  }, [customerOrders?.orders.edges, orderedBrands, recommendedProducts?.products])
+
+  return recommendedProduct
+}
+
+export function addRecommendedProduct(products: Product[], recommendedProduct: Product | undefined): void {
+  if (!products.some(product => product.recommended)) { // Check if no product has been recommended already
+    const result = products.findIndex(product => { return product.id === recommendedProduct?.id }) // Check if recommended product already exists in list of products
+    if (result != -1) { products.splice(result, 1) } // If it exists remove it
+    if (recommendedProduct) { products.unshift(recommendedProduct) } // Add recommended product
   }
 }
